Relocate abstract command base to cloud package
