+++ velocity/src/main/java/net/crystalixs/core/velocity/CorePlugin.java
@@ -10,9 +10,9 @@ import com.velocitypowered.api.proxy.Player;
 import com.velocitypowered.api.proxy.ProxyServer;
 import jakarta.inject.Inject;
 import net.crystalixs.core.common.config.ObjectMapperFactory;
-import net.crystalixs.core.velocity.command.VelocityCommandManagerTypeLiteral;
-import net.crystalixs.core.velocity.command.VelocityCommandSource;
-import net.crystalixs.core.velocity.command.VelocityPlayerCommandSource;
+import net.crystalixs.core.velocity.command.cloud.VelocityCommandManagerTypeLiteral;
+import net.crystalixs.core.velocity.command.cloud.VelocityCommandSource;
+import net.crystalixs.core.velocity.command.cloud.VelocityPlayerCommandSource;
 import net.crystalixs.core.velocity.config.VelocityConfig;
 import net.crystalixs.core.velocity.config.VelocityConfigLoader;
 import net.crystalixs.core.velocity.config.jackson.JacksonVelocity;
