+++ velocity/src/main/java/net/crystalixs/core/velocity/command/cloud/VelocityCommandSource.java
@@ -1,4 +1,4 @@
-package net.crystalixs.core.velocity.command;
+package net.crystalixs.core.velocity.command.cloud;
 
 import com.velocitypowered.api.command.CommandSource;
 import net.crystalixs.core.common.command.AbstractCommandSource;
